# backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Architecture Platform API
  description: |
    # Architecture Platform API
    
    Complete API documentation for the Architecture Platform microservices.
    
    ## Authentication
    All API requests require a JWT token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited per service:
    - User Service: 60 requests/minute
    - Project Service: 100 requests/minute
    - Property Service: 200 requests/minute
    - Version Service: 100 requests/minute
    - Search Service: 300 requests/minute
    - Collaboration Service: 500 requests/minute
    - Geometry Service: 200 requests/minute
    - Script Service: 50 requests/minute
    - File Service: 100 requests/minute
    - Notification Service: 200 requests/minute
    - Analytics Service: 150 requests/minute
  version: 1.0.0
  contact:
    name: Architecture Platform Team
    email: api@archplatform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.archplatform.com/api/v1
    description: Production server
  - url: https://api-dev.archplatform.com/api/v1
    description: Development server
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List projects
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, archived, deleted]
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProjects'

    post:
      tags: [Projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /projects/{id}:
    get:
      tags: [Projects]
      summary: Get project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

  # ==========================================
  # Properties
  # ==========================================
...
  /search:
    get:
      tags: [Search]
      summary: Search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [project, user, file, all]
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
  /search/suggestions:
    get:
      tags: [Search]
      summary: Search suggestions
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search suggestions

  # ==========================================
  # Files
  # ==========================================
...
  /scripts:
    post:
      tags: [Scripts]
      summary: Create script
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Script'
      responses:
        '201':
          description: Script created
  /scripts/{id}/execute:
    post:
      tags: [Scripts]
      summary: Execute script
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                inputs:
                  type: object
      responses:
        '200':
          description: Execution result
          content:
            application/json:
              schema:
                type: object
                properties:
                  output:
                    type: string
                  execution_time_ms:
                    type: integer

  # ==========================================
  # Notifications
  # ==========================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
  schemas:
...
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        location:
          type: string
        metadata:
          type: object

    # Responses
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        owner_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, archived, deleted]
        member_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PaginatedProjects:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        page:
          type: integer
        page_size:
          type: integer
        total_count:
          type: integer
        total_pages:
          type: integer
    SearchResults:
      type: object
      properties:
        query:
          type: string
        total:
          type: integer
        hits:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              type:
                type: string
              title:
                type: string
              highlight:
                type: string

    # Error
...